Move helpers; clean up outdated logic
