+++ graph/src/index.ts
@@ -1,18 +1,15 @@
-import { ResourceEffects } from "./graph.js";
-import {
-  Cond,
-  UnwrapCond,
-  combineCond,
-  condIsLeaf,
-  evaluateCond,
-  flattenCond,
-  flattenSoloCond,
-  mapCond,
-  mapCondArr,
-  prettyPrintEdgeCondition,
-} from "./cond.js";
+import { Cond, combineCond } from "./cond.js";
 import { states as allStates, edges as allEdges, resources } from "./state.js";
 import _ from "lodash";
+import {
+  EdgeConditionWithResource,
+  HorizonEdgeCondition,
+  backpropagateCondition,
+  edgeConditionIsValid,
+  prettyPrint,
+  simplifyHorizonEdgeCond,
+  verifyCond,
+} from "./stateCond.js";
 
 const parseGraphValidity = (s: typeof allStates, e: typeof allEdges) => {
   if (s.length !== _.uniqBy(s, "id").length) {
@@ -27,86 +24,10 @@ const parseGraphValidity = (s: typeof allStates, e: typeof allEdges) => {
   // should URLs and startingness be separate? maybe URLs as groups?x
 };
 
-const findAssociatedEdges = (
-  s: (typeof allStates)[number],
-  edges: typeof allEdges
-) => edges.filter((edge) => edge.from === s.id);
-
-const findState = (
-  id: (typeof allStates)[number]["id"],
-  states: typeof allStates
-  // id is guaranteed to exist
-) => states.find((s) => s.id === id)!;
-
 interface Step {
   edgeName: string;
   type: "prep" | "action" | "cleanup";
 }
-
-const wasEdgeTraversed = (edgeName: string, steps: Step[]) =>
-  steps.some((step) => step.edgeName === edgeName);
-
-type VisitedStatesRecord = Partial<
-  Record<(typeof allStates)[number]["id"], boolean>
->;
-
-// recursively traverses the graph using DFS, scheduling actions and cleanups.
-// avoids retreading visited states
-const traverseDFS = (
-  e: typeof allEdges,
-  state: (typeof allStates)[number],
-  visitedStates: VisitedStatesRecord = {}
-): { steps: Step[]; visitedStates: VisitedStatesRecord } => {
-  const edges = findAssociatedEdges(state, e);
-  let currVisitedStates: Partial<
-    Record<(typeof allStates)[number]["id"], boolean>
-  > = { ...visitedStates, [state.id]: true };
-
-  const currSteps = _.flatMap(edges, (edge) => {
-    if (
-      // TODO every edge should be traversed once
-      // wasEdgeTraversed(edge.name, steps) &&
-      currVisitedStates[edge.to]
-    )
-      return []; // already visited, skip
-    const nextState = findState(edge.to, allStates);
-
-    const result = traverseDFS(e, nextState, currVisitedStates);
-
-    const toReturn = [
-      { edgeName: edge.name, type: "action" as const },
-      ...result.steps,
-      { edgeName: edge.name, type: "cleanup" as const },
-    ];
-
-    // track visited states per sibling
-    _.forEach(
-      toReturn.filter((step) => step.type === "action"),
-      (step) => {
-        const edge = edges.find((e) => e.name === step.edgeName);
-        currVisitedStates = {
-          ...currVisitedStates,
-          ...result.visitedStates,
-          ...(edge ? { [edge.to]: true } : {}),
-        };
-      }
-    );
-
-    return toReturn;
-  });
-
-  return { steps: currSteps, visitedStates: currVisitedStates };
-};
-
-const simpleScheduler = (
-  states: typeof allStates,
-  edges: typeof allEdges
-): Step[] =>
-  _.flatMap(edges, (edge) => [
-    { edgeName: edge.name, type: "prep" },
-    { edgeName: edge.name, type: "action" },
-    { edgeName: edge.name, type: "cleanup" },
-  ]);
 
 const getStartingStates = (states: typeof allStates) =>
   states.filter((s) => !_.isNil(s.url));
@@ -149,266 +70,7 @@ const traversabilityCheck = (
   return _.every(stateDict);
 };
 
-type EdgeConditionWithResource = UnwrapCond<
-  NonNullable<(typeof allEdges)[number]["condition"]>
->;
-const prettyPrint = (condition: Cond<EdgeConditionWithResource | boolean>) =>
-  prettyPrintEdgeCondition(condition, (c) =>
-    _.isBoolean(c)
-      ? c.toString()
-      : `${c.operator === "gt" ? ">" : "<"} ${c.value} ${c.resource}`
-  );
-
-const backpropagateCondition = (
-  cond: Cond<EdgeConditionWithResource | boolean>,
-  resourceEffects:
-    | ResourceEffects<EdgeConditionWithResource["resource"]>
-    | undefined
-): Cond<EdgeConditionWithResource | boolean> => {
-  return mapCond(cond, (c) => {
-    if (_.isBoolean(c) || _.isNil(resourceEffects)) return c;
-    const resourceEffect: number | undefined = resourceEffects[c.resource];
-    if (_.isNil(resourceEffect)) return c;
-    return { ...c, value: c.value - resourceEffect };
-  });
-};
-
-const verifyCond = (
-  value: number,
-  resource: EdgeConditionWithResource["resource"],
-  cond: Cond<EdgeConditionWithResource | boolean>
-): boolean =>
-  evaluateCond(cond, (c) => {
-    if (_.isBoolean(c)) return c;
-    if (c.resource === resource) {
-      if (c.operator === "lt") return value < c.value;
-      if (c.operator === "gt") return value > c.value;
-    }
-    return true;
-  });
-
-const edgeConditionIsValid = (
-  cond: Cond<EdgeConditionWithResource | boolean>
-): boolean => {
-  const flattenedConditions: (EdgeConditionWithResource | boolean)[] =
-    flattenCond(cond);
-
-  const sets: Partial<
-    Record<EdgeConditionWithResource["resource"], Set<number>>
-  > = {};
-  const addToSet = (
-    resource: EdgeConditionWithResource["resource"],
-    value: number
-  ) => {
-    if (!sets[resource]) sets[resource] = new Set();
-    sets[resource]?.add(value);
-  };
-
-  _.forEach(flattenedConditions, (c) => {
-    if (_.isBoolean(c)) return;
-    addToSet(c.resource, c.value);
-  });
-
-  try {
-    _.forEach(sets, (set, resource) => {
-      if (_.isNil(set) || set.size <= 0) return true; // skip empty sets
-      const orderedValues = _.sortBy(Array.from(set));
-      const shiftedDownByOne = _.map(orderedValues, (v) => v - 1);
-      // appease TS with ?? 0. orderedValues always contains at least 1 value
-      const finalShiftedUpOne = (_.last(orderedValues) ?? 0) + 1;
-      const testValues = _.uniq([
-        ...orderedValues,
-        ...shiftedDownByOne,
-        finalShiftedUpOne,
-      ]);
-
-      if (
-        _.every(
-          testValues,
-          (v) =>
-            !verifyCond(
-              v,
-              resource as EdgeConditionWithResource["resource"],
-              cond
-            )
-        )
-      ) {
-        throw new Error("Invalid condition");
-      }
-    });
-  } catch (e) {
-    console.log(cond);
-    console.error(e);
-    return false;
-  }
-
-  return true;
-};
-
 const iterLimit = 10;
-
-type HorizonEdgeCondition =
-  | Cond<EdgeConditionWithResource | boolean>
-  | undefined;
-
-// TODO test
-// TODO determine numeric ranges and reconstruct a cond
-const simplifyHorizonEdgeCond = (
-  cond: HorizonEdgeCondition
-): HorizonEdgeCondition => {
-  if (_.isNil(cond)) return undefined;
-
-  const combinedCond = combineCond(cond);
-
-  const mappedCond = {
-    _and: mapCondArr(
-      [combinedCond],
-      (conds, parentType) => {
-        if (parentType === "and") {
-          if (_.some(conds, (c) => c === false)) return [false];
-          // defines the highest and lowest valid value boundary for each
-          // resource. if a value is defined, values *beyond* that value are
-          // valid. if a resource is not defined, all values are valid.
-          const bounds: Partial<
-            Record<
-              EdgeConditionWithResource["resource"],
-              { highest: number; lowest: number }
-            >
-          > = {};
-          const setBoundary = (
-            resource: EdgeConditionWithResource["resource"],
-            type: "high" | "low",
-            value: number
-          ) => {
-            const boundary = _.cloneDeep(bounds[resource]) ?? {
-              highest: Infinity,
-              lowest: -Infinity,
-            };
-            if (type === "high") {
-              boundary.highest = Math.min(boundary.highest, value);
-            } else {
-              boundary.lowest = Math.max(boundary.lowest, value);
-            }
-            bounds[resource] = boundary;
-            return boundary;
-          };
-          try {
-            _.forEach(conds, (c) => {
-              if (_.isBoolean(c)) return true; // continue
-              if (!condIsLeaf(c)) return true; // continue
-              const newBoundary = setBoundary(
-                c.resource,
-                c.operator === "lt" ? "high" : "low",
-                c.value
-              );
-              if (newBoundary.highest <= newBoundary.lowest)
-                throw new Error("Invalid cond");
-            });
-          } catch {
-            return [false];
-          }
-
-          const newConds: typeof conds = _.compact(
-            _.flatMap(
-              bounds,
-              (
-                boundary: (typeof bounds)[keyof typeof bounds],
-                resource: keyof typeof bounds
-              ) => {
-                if (_.isNil(boundary)) return undefined;
-
-                const lowerBound =
-                  boundary.lowest > -Infinity
-                    ? { resource, value: boundary.lowest, operator: "gt" }
-                    : undefined;
-                const upperBound =
-                  boundary.highest < Infinity
-                    ? { resource, value: boundary.highest, operator: "lt" }
-                    : undefined;
-
-                return [lowerBound, upperBound];
-              }
-            )
-          );
-
-          return [..._.filter(conds, (c) => !condIsLeaf(c)), ...newConds];
-        }
-
-        if (_.some(conds, (c) => c === true)) return [true];
-        // defines the highest and lowest valid value boundary for each
-        // resource. all values *outside* of that boundary are valid. if a
-        // resource is not defined, no values are valid.
-        const bounds: Partial<
-          Record<
-            EdgeConditionWithResource["resource"],
-            { highest: number; lowest: number }
-          >
-        > = {};
-        const setBoundary = (
-          resource: EdgeConditionWithResource["resource"],
-          type: "high" | "low",
-          value: number
-        ) => {
-          const boundary = _.cloneDeep(bounds[resource]) ?? {
-            highest: Infinity,
-            lowest: -Infinity,
-          };
-          if (type === "high") {
-            boundary.highest = Math.min(boundary.highest, value);
-          } else {
-            boundary.lowest = Math.max(boundary.lowest, value);
-          }
-          bounds[resource] = boundary;
-          return boundary;
-        };
-        try {
-          _.forEach(conds, (c) => {
-            if (_.isBoolean(c)) return true; // continue
-            if (!condIsLeaf(c)) return true; // continue
-            const newBoundary = setBoundary(
-              c.resource,
-              c.operator === "lt" ? "high" : "low",
-              c.value
-            );
-            if (newBoundary.highest <= newBoundary.lowest)
-              throw new Error("Tautology found");
-          });
-        } catch {
-          return [true];
-        }
-
-        const newConds: typeof conds = _.compact(
-          _.flatMap(
-            bounds,
-            (
-              boundary: (typeof bounds)[keyof typeof bounds],
-              resource: keyof typeof bounds
-            ) => {
-              if (_.isNil(boundary)) return undefined;
-
-              const lowerBound =
-                boundary.lowest > -Infinity
-                  ? { resource, value: boundary.lowest, operator: "lt" }
-                  : undefined;
-              const upperBound =
-                boundary.highest < Infinity
-                  ? { resource, value: boundary.highest, operator: "gt" }
-                  : undefined;
-
-              return [lowerBound, upperBound];
-            }
-          )
-        );
-
-        return [..._.filter(conds, (c) => !condIsLeaf(c)), ...newConds];
-      },
-      "and",
-      true
-    ),
-  };
-
-  return flattenSoloCond(mappedCond);
-};
 
 // Check if all conditional edges have some route from a starting state that
 // satisfies their conditions
@@ -845,9 +507,6 @@ const runScheduler = (
   edges: typeof allEdges
 ): Step[] => {
   console.log("Running Scheduler");
-  // for now, run every edge's prep, action, and cleanup
-  // return simpleScheduler(states, edges);
-
   // TODO new scheduler
   // determine starting states from urls
   // create implied edges from starting states to all other states
@@ -878,7 +537,6 @@ const runScheduler = (
 
   // TODO
   return [];
-  // console.log(traverseDFS(edges, states[0]!));
 };
 
 const runSteps = (steps: Step[], edges: typeof allEdges) => {
